+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+- Added option to accumulate gradients over multiple batches
+
 ## [0.1.6] - 2024-10-27
 - Minor change to Pipfile
 - Minor change to a unit test setup